requirements.txt
Flask==2.3.3
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
